[codeReview] fixing up location user controller and test

+++ Hack/Backend/controllers/LocationUserController.js
@@ -23,6 +23,12 @@ exports.post_location_user = async (req, res) => {
   if (!req.body.longitude || !req.body.latitude) {
     return res.status(400).send("Coordinates not given");
   }
+  if (!req.body.user_id) {
+    return res.status(400).send("User ID not given");
+  }
+  if (!req.body.TTL) {
+    return res.status(400).send("TTL not given");
+  }
   let latitude = req.body.latitude;
   let longitude = req.body.longitude;
 
@@ -104,6 +110,9 @@ exports.post_location_user = async (req, res) => {
       check.TTL = Number(TTL)
       check.probability = probability
       await check.save()
+      user.latest_location = check.location_id
+      user.latest_location_user = check._id
+      await user.save()
       return res.send(check);
     }
     location_user.probability = probability;
@@ -113,8 +122,7 @@ exports.post_location_user = async (req, res) => {
     await user.save();
     return res.send(location_user);
   } catch (err) {
-    console.log(err);
-    return res.status(500).send(err.toString());
+    return res.status(204).send("Location User could not be posted");
   }
 }
 
@@ -125,15 +133,17 @@ exports.get_location_user_by_id = async (req, res) => {
   }else{
     var LocationUser = LocationUserModels.LocationUser;
   }
-
+  if(!req.params.id){
+    return res.status(400).send("User Location ID not provided");
+  }
   try {
     const locationUser = await LocationUser.findById(req.params.id);
     if (!locationUser) {
-      return res.status(500).send("User Location not found");
+      return res.status(204).send("User Location not found");
     }
     res.send(locationUser);
   } catch (err) {
-    res.status(500).send(err.toString());
+    res.status(204).send("User Location not found");
   }
 }
 
@@ -144,16 +154,16 @@ exports.get_by_location_id = async (req, res) => {
   }else{
     var LocationUser = LocationUserModels.LocationUser;
   }
+  if(!req.params.location_id){
+    return res.status(400).send("Location ID not provided");
+  }
   try {
     const results = await LocationUser.find({
       location_id: { $eq: req.params.location_id },
     });
-    if (!results || results.length < 1) {
-      return res.status(500).send("No User Locations found.");
-    }
     res.send(results);
   } catch (err) {
-    res.status(500).send(err.toString());
+    return res.status(204).send("User Locations filtered by Location not found");
   }
 };
 
@@ -164,12 +174,11 @@ exports.get_all_location_users = async (req, res) => {
   }else{
     var LocationUser = LocationUserModels.LocationUser;
   }
-
   const results = await LocationUser.find({});
   try {
     res.send(results);
   } catch (err) {
-    res.status(500).send(err.toString());
+    res.status(204).send("No Location Users not found");
   }
 };
 
@@ -180,17 +189,16 @@ exports.get_by_user_id = async (req, res) => {
   }else{
     var LocationUser = LocationUserModels.LocationUser;
   }
-
+  if(!req.params.user_id){
+    return res.status(400).send("User ID not provided");
+  }
   try {
     const results = await LocationUser.find({
       user_id: { $eq: req.params.user_id },
     });
-    if (!results || results.length < 1) {
-      return res.status(500).send("No User Locations found.");
-    }
     res.send(results);
   } catch (err) {
-    res.status(500).send(err.toString());
+    return res.status(204).send("User Locations filtered by User not found");
   }
 };
 
@@ -201,15 +209,17 @@ exports.delete_location_user = async (req, res) => {
   }else{
     var LocationUser = LocationUserModels.LocationUser;
   }
-
+  if(!req.body._id){
+    return res.status(400).send("User Location ID not given");
+  }
   try {
     const location_user = await LocationUser.findByIdAndDelete(req.body._id);
     if (!location_user) {
-      return res.status(404).send("No item found");
+      return res.status(204).send("User Location Not Found");
     }
     res.status(201).send(location_user);
   } catch (err) {
-    res.status(500).send(err.toString());
+    res.status(204).send("Location User could not be deleted");
   }
 };
 
@@ -220,17 +230,19 @@ exports.update_location_user = async (req, res) => {
   }else{
     var LocationUser = LocationUserModels.LocationUser;
   }
-
+  if(!req.body._id){
+    return res.status(400).send("User Location ID not given");
+  }
   try {
     let locationUser = await LocationUser.findById(req.body._id);
     if (!locationUser) {
-      return res.status(500).send("User Location Not Found");
+      return res.status(204).send("User Location Not Found");
     }
     locationUser.set(req.body);
     await locationUser.save();
     res.send(locationUser);
   } catch (err) {
-    res.status(500).send(err.toString());
+    res.status(204).send("Location User could not be updated");
   }
 };
